Implementation customer Service

+++ src/main/kotlin/br/com/dastec/gerenciasalao/controllers/requests/services/PostServiceRequest.kt
@@ -11,5 +11,5 @@ data class PostServiceRequest(
     @JsonAlias("id_category")
     val idCategory: Long,
 
-    val price: BigDecimal?
+    val price: Double?
 )
